requirements.txt
python-dotenv
discord.py
mariadb
sqlalchemy
alembic
pandas
matplotlib
